+++ android/app/src/main/java/tw/kgame/crapto1/Span.java
@@ -3,25 +3,25 @@ package tw.kgame.crapto1;
 import java.util.Arrays;
 
 public class Span {
-    int[] array;
+    long[] array;
     int offset;
     int length;
 
-    public Span(int[] array) {
+    public Span(long[] array) {
         this(array, 0, array.length);
     }
 
-    public Span(int[] array, int offset, int length) {
+    public Span(long[] array, int offset, int length) {
         this.array = array;
         this.offset = offset;
         this.length = length;
     }
 
-    public int get(int i) {
+    public long get(int i) {
         return array[i + offset];
     }
 
-    public void set(int i, int v) {
+    public void set(int i, long v) {
         array[i + offset] = v;
     }
 
@@ -35,7 +35,7 @@ public class Span {
 
     public int binarySearch() {
         int start = 0, stop = this.length - 1, mid;
-        int val = this.get(stop) & 0xff000000;
+        long val = this.get(stop) & 0xff000000;
         while (start != stop)
             if (this.get(start + (mid = (stop - start) >> 1)) > val)
                 stop = start + mid;
